app: Acceptable overlay performance
