+++ horture.cabal
@@ -29,6 +29,18 @@ library horture-lib
 
     -- LANGUAGE extensions used by modules in this package. other-extensions:
     build-depends:    base ^>=4.14.3.0
+                    , X11
+                    , text
+                    , OpenGL
+                    , mtl
+                    , GLFW-b
+                    , GLUtil
+                    , raw-strings-qq
+                    , bytestring
+                    , clock
+                    , deepseq
+                    , synchronous-channels
+                    , containers
     ghc-options:
                     -Wall
                     -fno-warn-name-shadowing
@@ -37,10 +49,12 @@ library horture-lib
 
 executable horture-exe
     hs-source-dirs:   app
+    -- other-modules:
     main-is:          Main.hs
     ghc-options:
                     -fno-warn-name-shadowing
                     -threaded
+                    -O2
     build-depends:    base ^>=4.14.3.0
                     , horture-lib
                     , X11
@@ -54,4 +68,6 @@ executable horture-exe
                     , clock
                     , deepseq
                     , synchronous-channels
+                    , containers
+                    , optparse-applicative
     default-language: Haskell2010
